Add files via upload
This is a simple command-line calculator built in Python as part of my early learning journey. It performs basic arithmetic operations like addition, subtraction, multiplication, and division based on user input.

I built this project to understand how functions, conditionals, and user input work in Python. It helped me practice clean coding, logic structuring, and error-free execution.

This calculator runs directly in the terminal and is perfect for beginners who want to understand how to handle basic inputs and outputs in Python.

Note: The Calculator does round to the nearest 4 d.p.